pyproject.toml
[project]
name = "lab01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pycryptodome>=3.21.0",
]
